Add user controller tests for redirect when not logged in
The previous tests passed even without the before_action filters in use.
A massive security hole. The before filter operates on a per action
basis. These tests hit edit and update actions with right kind of
requests and verify the flash and that the request is redirected to
login_path.