Add tests for users index for admin and non-admin
Modified setup to reflect admin and non-admin users. Test for admin
asserts that there are links to each user and delete links to all
non-admin users. Asserts that user count changes when a user is deleted
and that the server responds and redirects to users index.

Test for non-admin user makes sure that there are no delete links on the
page.