Add mailer test for account activation email
Adds activation token to our test user fixture before generating the
email with UserMailer. Checks the subject, target email address to match
user.email and from address. In the email body check that user.name is
present and correct with activation token and that the email address is
correcly escaped.