Add update action to password resets controller
First checks if the password is empty and adds an error message
accordingly. It's currently allowed by our User model.

Then if the User update is successful the session is reset, user is
logged in and a flash signaling success is rendered on the user profile
page.

Otherwise we re-render the edit page.

`user_params` method permits only password and -confirmation.