Add tests for profile page microposts view
`response.body` contains full HTML source of the page. When asserting a
match with micropost count it returns true if the count is correct
anywhere on the page.

`h1>img.gravatar` is a nesting syntax for `assert_select`. This checks
for an `img` tag with class `gravatar` inside `h1` heading tag.