Add partial for micropost feed
Ordered list with pagination.

This feed partial defers rendering to the micropost partial defined
earlier. Rails knows to call the micropost partial because each element
of `@feed_items` has class `Micropost`. This causes Rails to look for a
partial with the corresponding name in the views directory of the given
resource.