Add destroy action to microposts controller
Before action filter correct_user checks that current user is the owner
of the micropost. User is found by user association of micropost which
has an added security benefit of failing automatically by returning
`nil` if a user tries to delete another user's micropost. Redirects to
root if it doesn't.

Because users can delete microposts from either the profile page or the
home page, it's convenient to redirect back to the referring page. This
is related to `request.original_url`, and in this case is just the
previous url.

Two edge cases:
1) referring url is sometimes `nil` inside of tests, so we need to
   redirect to a default url when that is the case.
2) In the case of trying to delete a micropost from the home page after
   submitting an invalid micropost, the application sends a `POST`
   request to the `microposts_url`. If the micropost is invalid, the app
   re-renders the home page. If the user then tries to delete a
   micropost on that page, the referring url is `microposts_url` and a
   redirect back - a regular `GET` request - will fail because there is
   no route corresponding to `GET /microposts`.

   **I've added this route and the latter part of the conditional seems
   redundant** Emailed Michael to confirm this.