Add relationship relations to models
In `User` model we need the `Relationship`  model class name specified,
because in User has_many `:active_relationships` when the underlying model
is called `Relationship`.

Similarly in Relationship we need to specify `User` when in the table
they're called `follower` and `followed`.

The `foreign_key` is the user id and it's used to connect two database
tables.