Add forms for following and unfollowing a user
These use `form_with` to manipulate a Relationship mode object. Follow
sends a `POST` request to Relationships controller to `create` a
Relationship and unfollow sends a `DELETE` request to `destroy` a
Relationship. The actions themselves are still to be defined.

The follow form has but a button, but it still needs to send the
`followed_id` to the controller. It's accomplished with the
`hidden_field_tag` method, which produces HTML of the form `<input
id='followed_id' name='followed_id' type='hidden' value='3' />`