Add turbo stream templates for relationship actions
`turbo_stream.update` takes in a css id and replaces the corresponding
element with the result of the block passed erb to the method.

Clicking follow/unfollow sends a Turbo-stream request to the Rails
server.

If there isn't any code in the action to respond to a
Turbo-stream, the request is handled as a regular HTML request.

If there is code to respond to a Turbo-stream, Rails evaluates a Turbo
template with the same name as the corresponding action.

Turbo templates can use the `turbo_stream.update` method to replace the
contents of a particular HTML element with the result of evaluating a
block of erb passed to the template.

All this results in a responsive site with the speed of a JS SPA and the
convenience of doing the heavy lifting on the server with Rails.