Push the finding of followed ids into database
Previous feed code pulled all the followed users' ids into memory,
because the condition `IN (?)` just checks the inclusion in a set.

Solution is to push the finding of followed ids into the database using
a subselect. SQL is optimized for this kind of work.

First step is to use a hash-based syntax for conveniency. We're going to
want to use `user_id` a second time elsewhere.