Use SQL subselect to get following_ids
Internally this select for user 1 would look like this:

```
SELECT * FROM microposts
WHERE user_id IN (SELECT followed_id FROM relationships
                  WHERE follower_id = 1)
      OR user_id = 1
```

This subselect arranges for the set logic to be pushed into the
database, which is more efficient.

However this only pulls out the feed's microposts without the associated
users or attachments. As a result, the feed partial, which calls the
micropost partial, generates an extra database hit to find the user
corresponding to each micropost. The feed view itself ends up generating
an additional N queries to find each micropost's user. The same applies
to as many as N attached images.

This N + 1 query problem drags down an app's performance as the size of
the database grows.